documentation/poky-ref-manual:  Figures cleanup
I removed two figures from the figures directory:
cropped-yocto-project-bw.png and yocto-project-transp.png.
Both figures are relics and not used in the manual.  I also
altered the Makefile to pull the ss-sato.png file from the
figures directory instead of the screenshots directory.  I moved
this PNG file from the screenshots directory to the figures
directory so that all figures would be in the figures directory.
Finally, I updated the introduction.xml file so that the html
code to include the ss-sato.png file pulls it from the figures
directory and not the screenshots directory.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>