documentation/bsp-guide/bsp.xml: Updated /binary explanation
In the '1.1.3 Pre-built User Binaries' section it said that the
ADT and minimal images were kept in the optional
meta-<bsp_name>/binary directory.  Jianjun Xu pointed out that
in fact it is just the minimal and sato images here.  I also
confirmed with Tom as well.  I re-wrote this sentence to be
clearer and more accurate.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>