[BUGID#_1025] - documentation/Makefile: Cleaned up Makefile
[BUGID#_1025] - I added some conditionals to handle the
Yocto Project Quick Start case.  This manual does not have
a PDF version.  I put in tests for publishing and for a case
where a user might attempt to specifically generate a PDF
using 'make pdf'.  I also converted the version variable into
a command-line argument so we don't have to edit the make file
when a new release comes out.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>