documentation/yocto-project-qs/yocto-project-qs.xml: removed 5.0 references
Removed several instances to the 5.0 Bernard release.  I replaced
these with the 5.0.1 release.  I also re-wrote a paragraph that
instructs the user on where to find and download the most recent
YP tarball.  It used to point to the Index of /Downloads into the
"yocto-1.0" folder.  I now instruct the user to go to the
Yocto Project website and download the desired release from there.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>