documentation/poky-ref-manual/extendpoky.xml: fixed meta/classes/core-image ref
In the section that describes how to customize images using IMAGE_FEATURES
I had the file meta/classes/core-image.bbclass appearing in a paragraph.
For the Bernard release this file should be
meta/classes/poky-image.bbclass.  I have made the correction.
The correction is in the yocto-docs/bernard branch only.  It is also
correct on the yocto website as well.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>