documentation/poky-ref-manual/faq.xml: Updated Python FAQ entry
The entry previously stated that only Python 2.6 is required.  I
added Python 2.7 to that requirement.  I also changed the URLs that
point to our pre-built tarballs from which the user can get Python
2.6 to be in the yocto-1.0 download area.  They were pointing to
an area in the index of downloads that was up one level.  According
to Beth the official location is
http://autobuilder.yoctoproject.org/downloads/miscsupport/yocto-1.0- \
python-nativesdk.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>