documentation/yocto-project-qs/yocto-project-qs.xml: added libtool to packages
Turns out that the package libtool is necessary for working with ADT during
the installation process.  I need to mention this in the ADT manual but
it would be helpful if it was one of the packages in the QS along with
all the others.  I added libtool to the package requirements lists
for the three supported hosts.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>