documentation/adt-manual/adt-intro.xml: changed definition of ADT
The manual previously defined the "ADT" as the toolchain, the sysroot,
QEMU, and user-space tools.  This definition was creating some confusion
when it came to installing the "ADT".  Namely, some of the items you
don't even install, such as the user-space tools.  So I recast the
definition of ADT to mean the cross-toolchain and sysroot.  I cast the
other items as part of the overall development environment.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>