documentation/adt-manual/adt-prepare.xml:  Updates to ADT install steps
Significant re-write to every section in this chapter except the
"Setting Up the Environment" and "Kernels and Filesystem Images"
chapters.  I established some terms to communicate the build and
source trees and the build and source top-level directories for
YP.  These terms had not been consitently communicated.  I also
stepped through each of the three methods by which a user can
install ADT (and the cross-toolchain) and meticulously documented
the steps.  The result is a much more complete and clear set
of procedures.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>