documentation/dev-manual/dev-manual-cases.xml: partial - edits to clean terms up.
I made a pass through the chapter to try and lessen the confusion over
yp files in a git repo verses those unpacked from a tarball.  Also
tried to write the example for the 1.1 release but I don't have the
real software yet and can't supply actual names of files.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>