documentation/dev-manual/dev-manual-model.xml: referenced cloning kernel files.
In the flow section for the kernel modification section I added a new
step for getting a local copy of the linux yocto kernel files.  The section
indicates that you need a local Git repository of the files and then
references back to the Getting Started section for an example.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>