documentation/dev-manual/dev-manual-start.xml: Added to kernel source setup
I added instruction on how to set up a bare clone of the linux yocto
kernel files and then how to create a clone of the bare clone.
Finally added a part about setting up the meta-kernel-dev repository.
This setup should do it for those that want to modify the kernel and
not have to download source files from the upstream everytime they
make a change.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>