documentation/dev-manual/dev-manual-model.xml: re-write of the kernel dev overview
After getting the example in kenerl example appendix working I discovered
much that was wrong in this overview.  The overview attempts to generalize
all the situations for kernel modification.  These include code changes,
configuration changes, and the addition of new recipes.  It also includes
the situation where a developer wants to gather all the changes into a
separate layer for distribution and for when they want to push them up
the tree.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>