documentation/dev-manual/dev-manual-kernel-appendix.xml: Added kernel reconfigure example.
this example uses the CONFIG_SMP configuration.  By default, this configuration is
disabled.  The example shows how to use menuconfig and explains the .config file
role and how you can use it combined with menuconfig to manage multiple configuration
changes.

Thanks to Darren Hart for identifying this simple change to demonstrate how to
change the kernel configuration.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>