documentation/poky-ref-manual/faq.xml: Scrubbed for term Poky.
In general, the term Poky can be replaced by "the Yocto Project."
The term Poky can be used as a term for the Yocto Project build system.
I left it in place for a few areas where it was easy and would not
be confusing.

I also performed general rewrites for the FAQ.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>