YOCTO #910 - documentation/poky-ref-manual/extendpoky.xml: point to dev manual.
There is a section in the YP ref manual that gives reference type information
on adding a machine.  This type of information is detailed well in the
BSP example appendix in the development manual.  I don't want to completely
eliminate the information in the YP ref manual as is suggested by the
910 bug.  I think referencing it in the dev manual for now works.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>