documentation/dev-manual/dev-manual-model.xml: Edits to Modifying Kernel
Complete re-write of this section based on feedback from Dave Stewart.
Dave's comments centered around not being able to understand the overall
concepts of the Git repositories, the source areas on the host, etc.
I have added several illustrations and written around them to try and
better explain the kernel storage system in Git used by YP and how
that tranlates into host-resident code that the developer ultimately
uses and the Poky build system uses for the build.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>