documentation/bsp-guide/bsp.xml: Edits for Crownbay example in kernel recipes
Because the Crownbay uses both EMGD and non-EMGD statments in the
linux-yocto_3.0.bbappend file I had to do some explaining.  It turns out
you don't really need to just delete or comment out the non-applicable
statements.

I also adjusted some wording around the tuning file example.  This file
I assumed was in the conf directory of the meta-crownbay directory.
It is really in the meta/conf/machine/include directory.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>