documentation/dev-manual/dev-manual-newbie.xml: Suggestions Robert Berger
I made several edits here.  There is some mystery as to which changes
are actually part of this commit.  I did a 'git diff' and am only seeing
one small change.  If I later find they are not there I will delete the
file and replace with a back up I saved prior to submitting this commit.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>