documentation: Makefile corrections to support web server structure
The web server directory structure for the post 1.0 releases was changed.
Also, the creation of a new 1.0 area in the web structure was retroactively
created.  this broke the five make files for publishing documents to the web.

I fixed all five files so they now push to the 1.0 area only.  The fix included
hard-coding the 1.0 directory structure.  I also set them up to be a little more
generic.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>