documentation/poky-ref-manual: Enabling recipes with commercial license requirements
I created a new section in the "Technical Details" chapter of
the YP Reference Manual to deal with licensing.  I migrated the
existing section on tracking licenses from the "Extending YP"
chapter into this section.  The new section now includes that info
as well as the new information here.

Fixes [YOCTO #576]

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>