documentation: Variables for docs and QS variablized
Beth has created a scheme that uses a file named "poky.ent" that
contains a load of variables that can be used throughout the
documents for releases.  The variables can be set in the poky.ent
file and then sucked in during the make process to specify
a release.

This commit adds the new file and the changes to the YP QS manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>