documentation:  More information about DL_DIR added.
The information surrounding the central download directory was
pretty scant.  Jim Abernathy pointed this out in an email sent to
me.  As a start, I have added quite a bit more information to
the glossary description for DL_DIR.  Part of that added information
is a pointer to the FAQ entry that describes exactly how the
build system gets source files.  Then, in the YP Development Manual
in the section that introduces the build process, I added several
references to various variables that are set in the local.conf
file prior to kicking off the build.

Reported-by: Jim Abernathy <jim.abernathy@intel.com>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>