documentation/poky-ref-manual/ref-variables.xml: CONFFILES and FILES edits
Final edits to the new CONFFILES entry and an updated more robust
FILES entry in the glossary.  These two entries are good examples
of the depth of detail we should have for all glossary entries.
Kudos to Paul Eggleton for tirelessly working with me to get the
descriptions just right.

Reported-by: Paul Eggleton <paul.eggleton@linux.intel.com>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>