documentation/dev-manual/dev-manual-bsp-appendix.xml: tags/branches edits
Edits to the section describing how to set up the YP files before the
BSP example.  I am working with Joshua Lock on these to settle them
down.  The edits are trying to establish understanding between a
Git repository and a tarball... just what the user has a hold of
after getting the files by either method.

Reported-by: Joshua Lock <joshua.lock@intel.com>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>