documentation/kernel-manual/kernel-concepts.xml: New kernel section
This section is the first attempt at a new, high-level discussion
of Linux Yocto Kernel configuration.  The section is supposed to
introduce the concepts only and leave the "How to" stuff for other
areas of the documentation.  I have thrown this section out to the
Yocto Projects Discussion list for feedback.  It will change.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>