documentation/bsp-guide/bsp.xml: qualified the commit ID difference
In the example that shows the linux-yocto_3.0.bbappend file the
commit IDs are pretty much random and taken at the time I wrote
the section.  I have provided a qualifying statement before the
example to explain that the commit ID strings in the manual's
example could differ from those in the actual file in the YP
source repository.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>