documentation/dev-manual/dev-manual-common-tasks.xml: new section
Created a new section called "Configuring the Kernel."  This is the
how-to counterpart for the concepts part created in the kernel
manual.  The section consists of three sub-sections:
"Using menuconfig", "Creating Config Fragments", and
"Fine-tuning the Kernel Configuration File."  Primary source for
information on this section was Bruce Ashfield.  He has this section
for review.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>