documentation/dev-manual/dev-manual-common-tasks.xml: Combined sections
I combined the section on changing package source code and
the section on changing package source code with quilt into one
over-arching section named "Modifying Package Source Code."
The section on Quilt was sort of a stand-alone section but Paul
Eggleton suggested that it is a natural follow-on.  So I put the
two sections in as sub-sections.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>