documentation/dev-manual/dev-manual-common-tasks.xml: re-write to 4.4
Complete re-write to the "Modifying Temporary Source Code" section,
Section 4.4.  This strategy now comprises telling the user where
this temporary source code is, how to change it within a Quilt
workflow, and how to change it within a Git workflow.

I consulted with Paul Eggleton quite a bit to come to this conclusion
that the section needed more attention.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>