documentation/poky-ref-manual: Review edits on B and externalsrc.bbclass
I tried to implement some comments by Richard on the B glossary
entry and the externalsrc.bbclass section.  The comments were a bit
confusing to me but the gist of it was I mis-interpreted Richard's
comments on supporting separate directories for B and S.  I had
indicated that autotools-based recipes did this by default, which
was wrong.  What was confusing to me is where the comments should
be.  I ended up putting them in both areas of the reference manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>