documentation/dev-manual:  New Layer section and metadata link anchor
The change to the dev-manual-newbie.xml file was minor.  I added
a link anchor for the term "metadata."

The change to the dev-manual-common-tasks.xml was extensive.  I have
added an entirely new section on layers called "Understanding and
Using Layers."  The new material has several sub-sections.

Information was based on emails from Paul Eggleton.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>