documentation/dev-manual/dev-manual-start.xml: Wording change
Added a bit more information to the wording where it talks about
creating a Git Repository if you are going to contribute back
into YP.  This change was suggested by Robert P. J. Day and
was submitted to me in the form of a patch.  For reasons unknown
to me the patch Hunk failed even though I believe I had things
saved and applied correctly.  I had to resort to applying it by
hand.  I ended up doing a bit of re-writing on the suggested
change to clean it up.

Reported-by: Robert P. J. Day <rpjday@crashcourse.ca>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>