documentation: Updates for [YOCTO #1717].
Minor edits to the BSP Guide, the Development Manual, and the
Kernel Manual to address [YOCTO #1717], which noted confusion
over the kernel manual when a user was trying to find kernel
configuration information.

The fix for this bug involved several areas.  I had previously
created a new section called "Kernel Configuration" in the kernel
manual and "Configuring the Kernel" in the YP Development manual.
There were several rounds of feedback from Bruce Ashfield.

This commit represents the final touches on the three manuals to
address the bug.

Fixes [YOCTO #1717]

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>