documentation/adt-manual/style.css: Added new style for Note headings
The style for the note, caution, warning, and tip headings was a bit
over-bearing.  Feedback from the community indicated that those words
needed to be in a smaller, less noticeable font.  I created the style
so that the font's size is about half of what it was and I also changed
the color to match the other headings.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>