documentation/bsp-guide/bsp.xml: Edits to intro sections
Adding the new section "Customizing a Recipe for a BSP" has added
enough sections to this book such that the introductory area needed
some attention.  I pared down the intro text to basically just
cover the manual's organization and a bit about BSPs.  I moved
all the introductory common form stuff into the section that
just addresses the BSP form.  This is better organization for the
manual as it is evolving.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>