documentation:  Removed the 1.2.1 manual revision entry
Five manuals have a manual revision table.  I removed this entry because
this set of manuals is custom for the denzil 1.2 release.  I was running
into publishing problems for the 1.2 manual set as compared to the 1.2.1
manual set, which works off the denzil main branch.  The issue is that
the denzil-7.0 tag on the denzil branch is well below critical, post-
release commits for documentation.  Richard Purdie agreed to merge these
commits into the denzil branch so that I could build more up-to-date
1.2 manuals as compared to what is in the denzil poky tarball.  This is
not ideal but because of the way we end-gamed 1.2 much documentation
testing was not done.  Many errors were discovered.  This commit and
creation of the new '1.2-denzil' branch in my yocto-docs area should
settle the issue of trying to separately build 1.2 manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>