documentation/dev-manual:  Moved temp source code mod into dev model
The section "Modifying Temporary Source Code" that was in the chapter
for "Common Tasks" was moved to the "Common Development Models" chapter.
After discussion with Darren Hart, modifying temporary source code
is a valid develop model.  So, rather than isolate it in the common
tasks chapter it has been placed in the development model chapter,
which now has four types of development.

Changes were made to update the introductory text of all the concerned
areas of the manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>