documentation/dev-manual: Minor 1.1.2 updates
The second pass through the manual testing against the variables
put in during the first pass.  Minor changes and few links not
right.  One thing of note is the tarball used for the BSP appendix
example for Crown Bay.  The tarball I have in there is the
release 1.1 version with poky 6.0.0.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>