documentation/adt-manual/adt-prepare.xml: Re-worded toolchain generation
The section that describes how to generate the cross-toolchain into the
Yocto build tree erroneously referred to this process as installing the
toolchain.  Furthermore, it actually mentioned the creation of a
toolchain tarball, which is wrong.  I updated the section to more properly
refer to this procedure as generating the toolchain.  I also removed the
bit about the tarball creation.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>